% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fzero.R
\name{fzero}
\alias{fzero}
\title{A function used to get the root by the method of bisection}
\usage{
fzero(f, a, b, eps = 1e-05)
}
\arguments{
\item{f}{the equation which We need to find  the root}

\item{a}{The right endpoint of the interval at the root of the equation}

\item{b}{The right endpoint of the interval at the root of the equation}

\item{eps}{Iteration is stopped when both ends of the interval are less than eps}
}
\value{
a root of the equation
}
\description{
This function is used to calculate the root of a equation in a interval and you need promise the signs at both ends of the interval are opposite or it will fail
}
\examples{
f<-function(x) x^3-x-1
fzero(f, 1, 2, 1e-6)
\dontrun{
fzero(f, 0, 1, 1e-6)
}
}
