% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{getinterval}
\alias{getinterval}
\title{A function used to get interval estimation of linear regression parameters}
\usage{
getinterval(fm, alpha = 0.05)
}
\arguments{
\item{fm}{The regression model calculated by function lm}

\item{alpha}{alpha is the confidence level}
}
\value{
confidence intervals of linear regression parameters and the left(right) is the left(right) endpoint of the confidence interval
}
\description{
This function is used to calculate the interval estimation of linear regression parameters
}
\examples{
x<-c(0.10, 0.11, 0.12, 0.13, 0.14, 0.15,0.16, 0.17, 0.18, 0.20, 0.21, 0.23)
y<-c(42.0, 43.5, 45.0, 45.5, 45.0, 47.5,49.0, 53.0, 50.0, 55.0, 55.0, 60.0)
lm.sol<-lm(y ~ 1+x+I(x^2))
getinterval(lm.sol)
}
