% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliner.R
\name{outliner}
\alias{outliner}
\title{A function used to draw the outline graph of P variables}
\usage{
outliner(x)
}
\arguments{
\item{x}{is a matrix or a dataframe}
}
\value{
the outline graph of P variables
}
\description{
This function is used to draw the outline graph of P variables
}
\examples{
x <- replicate(4,sample(seq(60,100),10))
outliner(x)
}
