\name{StatComp18013-package}
\alias{StatComp18013-package}
\docType{package}
\title{
  This package provides several functions for and answers for 'Statistical Computing' course .
}
\description{
    This package provides several functions for and answers for 'Statistical Computing' course .The answers are in vignettes.
}
\author{
18013

Maintainer: Hong Zhang <zhangh@ustc.edu.cn>
}
\keyword{ package }
